Make code shield-friendly
The conditional part is placed at the end of vcl_recv, immediately after
the code handling shielding. The shielding code replaces the default
backend with a special backend for the shield pop if and only if the pop
is not the shield pop. Therefore if the backend is still the default
(and not this special backend), then either shielding is off or we are
already running in the shield pop. Also, since we override the backend
there is no need for the early return.

JIRA: PLAT-1427